+++ CHANGELOG.md
@@ -36,8 +36,8 @@ release.
 ## [Unreleased]
 
 ### Added
-
 - Added file creation permission verification before creating database and resolve database creation to specified mission list only if `mlist` is set [#75](https://github.com/DOI-USGS/SpiceQL/pull/75)
+- Added /searchForKernelsets REST endpoint [#76](https://github.com/DOI-USGS/SpiceQL/pull/76)
 
 ### Fixed
 - Fixed default SpiceQL REST URL [#63](https://github.com/DOI-USGS/SpiceQL/pull/63)